eat(client): applying UI - part 5
