+++ client/src/assets/jss/material-kit-react/components/buttonStyle.js
@@ -6,7 +6,7 @@ import {
   successColor,
   warningColor,
   dangerColor
-} from "assets/jss/material-kit-react.jsx";
+} from "assets/jss/material-kit-react";
 
 const buttonStyle = {
   button: {
