+++ client/src/assets/jss/material-kit-react/components/customDropdownStyle.js
@@ -12,7 +12,7 @@ import {
   dangerBoxShadow,
   roseColor,
   roseBoxShadow
-} from "assets/jss/material-kit-react.jsx";
+} from "assets/jss/material-kit-react";
 
 const customDropdownStyle = theme => ({
   popperClose: {
