+++ client/src/assets/jss/material-kit-react/components/headerLinksStyle.js
@@ -1,6 +1,6 @@
-import { defaultFont } from "assets/jss/material-kit-react.jsx";
+import { defaultFont } from "assets/jss/material-kit-react";
 
-import tooltip from "assets/jss/material-kit-react/tooltipsStyle.jsx";
+import tooltip from "assets/jss/material-kit-react/tooltipsStyle";
 
 const headerLinksStyle = theme => ({
   list: {
