+++ client/src/assets/jss/material-kit-react/components/headerStyle.js
@@ -10,7 +10,7 @@ import {
   transition,
   boxShadow,
   drawerWidth
-} from "assets/jss/material-kit-react.jsx";
+} from "assets/jss/material-kit-react";
 
 const headerStyle = {
   appBar: {
