+++ client/src/assets/jss/material-kit-react/components/parallaxStyle.js
