+++ client/src/assets/jss/material-kit-react/views/components.js
@@ -1,24 +1,21 @@
-import { container, title } from "assets/jss/material-kit-react.jsx";
+import { container } from "assets/jss/material-kit-react";
 
-const landingPageStyle = {
-  container: {
-    zIndex: "12",
+const componentsStyle = {
+  container,
+  brand: {
     color: "#FFFFFF",
-    ...container
+    textAlign: "left"
   },
   title: {
-    ...title,
+    fontSize: "4.2rem",
+    fontWeight: "600",
     display: "inline-block",
-    position: "relative",
-    marginTop: "30px",
-    minHeight: "32px",
-    color: "#FFFFFF",
-    textDecoration: "none"
+    position: "relative"
   },
   subtitle: {
     fontSize: "1.313rem",
     maxWidth: "500px",
-    margin: "10px auto 0"
+    margin: "10px 0 0"
   },
   main: {
     background: "#FFFFFF",
@@ -30,7 +27,13 @@ const landingPageStyle = {
     borderRadius: "6px",
     boxShadow:
       "0 16px 24px 2px rgba(0, 0, 0, 0.14), 0 6px 30px 5px rgba(0, 0, 0, 0.12), 0 8px 10px -5px rgba(0, 0, 0, 0.2)"
+  },
+  link: {
+    textDecoration: "none"
+  },
+  textCenter: {
+    textAlign: "center"
   }
 };
 
-export default landingPageStyle;
+export default componentsStyle;
