+++ client/src/components/Layout/CustomButtons/Button.js
@@ -10,7 +10,7 @@ import Button from "@material-ui/core/Button";
 
 // core components
 
-import buttonStyle from "assets/jss/material-kit-react/components/buttonStyle.jsx";
+import buttonStyle from "assets/jss/material-kit-react/components/buttonStyle";
 
 function RegularButton({ ...props }) {
   const {
