+++ client/src/components/Layout/CustomDropdown/CustomDropdown.js
@@ -16,9 +16,9 @@ import Icon from "@material-ui/core/Icon";
 import Popper from "@material-ui/core/Popper";
 
 // core components
-import Button from "../CustomButtons/Button.jsx";
+import Button from "../CustomButtons/Button";
 
-import customDropdownStyle from "assets/jss/material-kit-react/components/customDropdownStyle.jsx";
+import customDropdownStyle from "assets/jss/material-kit-react/components/customDropdownStyle";
 
 class CustomDropdown extends React.Component {
   constructor(props) {
