+++ client/src/components/Layout/Header/Header.js
@@ -14,7 +14,7 @@ import Drawer from "@material-ui/core/Drawer";
 // @material-ui/icons
 import Menu from "@material-ui/icons/Menu";
 // core components
-import headerStyle from "assets/jss/material-kit-react/components/headerStyle.jsx";
+import headerStyle from "assets/jss/material-kit-react/components/headerStyle";
 
 class Header extends React.Component {
   constructor(props) {
